% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print.pmxpartab}
\alias{print.pmxpartab}
\title{Print \code{pmxpartab} object.}
\usage{
\method{print}{pmxpartab}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link{pmxpartab}}.}

\item{...}{Further arguments passed on to other \code{print} methods.}
}
\value{
Returns \code{x} invisibly.
}
\description{
Print \code{pmxpartab} object.
}
\details{
In an interactive context, the rendered table will be displayed in
a web browser. Otherwise, the HTML code will be printed as text.
}
