% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nm_output.R
\name{read_meta}
\alias{read_meta}
\title{Read meta information from a YAML file}
\usage{
read_meta(meta.file = "meta.yaml")
}
\arguments{
\item{meta.file}{The name of a YAML file to be read.}
}
\value{
A \code{list} if the file exists, otherwise \code{NULL}.
}
\description{
Read meta information from a YAML file
}
