% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pmxpartab}
\alias{pmxpartab}
\title{Generate a parameter estimates table in HTML}
\usage{
pmxpartab(
  parframe,
  columns = c(est = "Estimate", rse = "RSE\%", ci95 = "95\%CI", shrinkage =
    "Shrinkage"),
  sections = TRUE,
  section.labels = c(Structural = "Typical Values", CovariateEffect =
    "Covariate Effects", RUV = "Residual Error", IIV = "Between Subject Variability", IOV
    = "Inter-Occasion Variability"),
  show.fixed.to.zero = F,
  na = "n/a",
  digits = 3
)
}
\description{
Generate a parameter estimates table in HTML
}
\examples{
outputs <- list(
    th = list(CL = 0.482334, VC = 0.0592686),
    om = list(nCL = 0.315414, nVC = 0.536025),
    sg = list(ERRP = 0.0508497),
    se = list(
        th = list(CL = 0.0138646, VC = 0.00555121),
        om = list(nCL = 0.0188891, nVC = 0.0900352),
        sg = list(ERRP = 0.00182851)),
    fixed = list(
        th = list(CL = FALSE, VC = FALSE),
        om = list(nCL = FALSE, nVC = FALSE),
        sg = list(ERRP = FALSE)),
    shrinkage = list(nCL = 9.54556, nVC = 47.8771))

meta <- list(
    parameters = list(
        list(name="CL",   label="Clearance",    units="L/h", type="Structural"),
        list(name="VC",   label="Volume",       units="L",   trans="exp", type="Structural"),
        list(name="nCL",  label="On Clearance",              trans="SD (CV\%)", type="IIV"),
        list(name="nVC",  label="On Volume",                 type="IIV"),
        list(name="ERRP", label="Proportional Error",        units="\%", trans="\%", type="RUV")))

parframe(outputs, meta)

pmxpartab(parframe(outputs, meta),
    columns=c(est="Estimate", rse="RSE\%", ci95="95\%CI", shrinkage="Shrinkage"))
}
